Add support for HLG HDR in AV1
This checks for HLG (ARIB STD-B67) support in AV1, which is often
supported on HDR TVs. The Cast SDK almost always is able to decode this
and tonemap it anyways.

A small change also edits the VP9 codec string to include the color
range enforcement flag. According to the spec, this is technically
required in the MIME string according to the spec, but in most browsers,
this is not enforced.

```
The first four components are required; everything from CC
(chroma subsampling) onward is optional, but all or nothing.
```